Emit a friendly message when the installer type is wrong.
See https://github.com/felddy/foundryvtt-docker/issues/282 for
details. The short version is that it's a common error when using
FOUNDRY_RELEASE_URL to download the windows installer. Previously
a confusing message was emitted by the unzip command when that
happened. This gives a message that should help beginners understand
their mistake in choosing the wrong installer type.